Our Custom Python scripts for Data Cleaning
'dupecheck.py' and 'resizing.py' were both used in the Data Cleaning process to check for duplicate images and standardize the datasets. The other 3 scripts were used for Data Visualization, plotting histograms of the pixel intensities (and aggregate pixel intensities)  and a bar graph to count the number of images in each dataset class.