Original and Generated Split Datasets
ProjectDatasets.zip: Original dataset that is split into the 4 classes Angry, Focused, Happy, and Neutral. The subdirectories are not divided further than that and each class contains at least 500 images. This dataset was used with the datasplitter.py script to generate the GeneratedSplitDataset which contains the split of training, validation, and testing sets of each class.

GeneratedSplitDataset.zip: This is the automatically generated dataset that contains the split of each classes' images into training, validation, and testing sets using our datasetsplitter.py script. Training sets contain 70% of their respective classes' images, validation and testing sets contain 15% of their respective classes' images each. This dataset was used in the training, validation, testing, and evaluation of our main CNN model and the 2 distinct variants.